requirements.txt
python-telegram-bot==13.7
requests==2.26.0
num2words==0.5.10
openai==0.27.0
python-dotenv==0.19.2
